Move the remote_configuration under cbcontainers and controllers
