+++ cbcontainers/remote_configuration/validation_test.go
@@ -7,8 +7,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	cbcontainersv1 "github.com/vmware/cbcontainers-operator/api/v1"
 	"github.com/vmware/cbcontainers-operator/cbcontainers/models"
-	"github.com/vmware/cbcontainers-operator/remote_configuration"
-	"github.com/vmware/cbcontainers-operator/remote_configuration/mocks"
+	"github.com/vmware/cbcontainers-operator/cbcontainers/remote_configuration"
+	"github.com/vmware/cbcontainers-operator/cbcontainers/remote_configuration/mocks"
 	"testing"
 )
